Delete deprecated functions/classes and add comments TokenMap switch map -> unordered_map
