+++ src/FileReader/FileReader.cpp
@@ -1,10 +1,14 @@
 #include <fstream>
-#include "FileReader.h"
 #include <iostream>
 #include <string>
 #include <sstream>
 
-std::string FileReader::ReadFile(const std::string &fileName) {
+/**
+ * Read file from specific folder
+ * @param fileName : Path of input file
+ * @return One line string with all text in file
+ */
+std::string static ReadFile(const std::string &fileName) {
     std::ifstream t(fileName);
     std::stringstream buffer;
     buffer << t.rdbuf();
