--- src/FileReader/FileReader.h
@@ -1,6 +1,0 @@
-#include <string>
-
-class FileReader {
-public:
-    static std::string ReadFile(const std::string &fileName);
-};
