--- src/Tokens/PreprocessedToken.cpp
@@ -1,5 +1,0 @@
-#include "PreprocessedToken.h"
-
-std::string PreprocessedToken::toString() const {
-    return value;
-}
