+++ src/Tokens/TokenNameMap.h
@@ -1,9 +1,14 @@
 #pragma once
 
 #include <string>
-#include <map>
+#include <unordered_map>
 #include "TokenCode.h"
-
+/**
+ * Class TokenMap is class which stores map \b TokenCode -> \b String. Used for demonstration for now.
+ * \Pattern
+ * The class uses the singleton template as the \b TokenMap \b class
+ * @see TokenMap.h
+ */
 class TokenNameMap {
 private:
     TokenNameMap(){GenerateMap();};
@@ -14,6 +19,6 @@ public:
         static TokenNameMap instance;
         return instance;
     }
-    std::map<TokenCode, std::string> tokenMap;
+    std::unordered_map<TokenCode, std::string> tokenMap;
     void GenerateMap();
 };
