+++ src/main.cpp
@@ -1,15 +1,21 @@
 #include <iostream>
-#include "FileReader/FileReader.h"
+#include "FileReader/FileReader.cpp"
 #include "LexicalAnalysis/Parser.h"
 #include "LexicalAnalysis/Demonstration.h"
 
+/**
+ Compilers Construction, team A
+ Implementing Compiler for 'Imperative toy language' on C++ to target code C
+ @authors Anna Kopeikina, Danila Korneenko, Tasha Sakovets, Yaroslav Kim, Dmitriy Pirozhenko
+ @version \b 1.1 \e 21/02/2023
+ **/
 int main(int argc, char *argv[]) {
 
-    std::string fileText = FileReader::ReadFile(
+    std::string fileText = ReadFile(
             R"(../../ILanguageCodeSnippets/ArraySum.i)");
     std::cout << fileText << "\n\n";
 
     Parser parser = Parser();
-    Demonstration::DemonstrateLexer(parser.GetLexicalAnalysisTokens(fileText));
+    DemonstrateLexer(parser.GetLexicalAnalysisTokens(fileText));
     return 0;
 }
